{"types":"C","runs":34}

+++ ConsoleSample.java
@@ -2,8 +2,8 @@ import acm.program.ConsoleProgram;
 
 public class ConsoleSample extends ConsoleProgram {
 	public void run() {
-	int n = readInt("n :");
-	int m = readInt ("m :");
+	double n = readInt("n :");
+	double m = readInt ("m :");
 	double average = (double) ((n+m)/2);
 	println (average);
 }
